+++ ElvUI/Modules/Skins/Blizzard/Inspect.lua
@@ -15,7 +15,6 @@ function S:Blizzard_InspectUI()
 
 	local InspectFrame = _G.InspectFrame
 	S:HandleFrame(InspectFrame, true, nil, 11, -12, -32, 76)
-
 	S:HandleCloseButton(_G.InspectFrameCloseButton, InspectFrame.backdrop)
 
 	for i = 1, #_G.INSPECTFRAME_SUBFRAMES do